+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADGroupLifecyclePolicy/MSFT_AADGroupLifecyclePolicy.psm1
@@ -203,34 +203,35 @@ function Set-TargetResource
     $null = New-M365DSCConnection -Workload 'MicrosoftGraph' -InboundParameters $PSBoundParameters
 
     $currentPolicy = Get-TargetResource @PSBoundParameters
+    $boundParameters = Remove-M365DSCAuthenticationParameter -BoundParameters $PSBoundParameters
 
     if ($Ensure -eq 'Present' -and $currentPolicy.Ensure -eq 'Absent')
     {
         Write-Verbose -Message "The Group Lifecycle Policy should exist but it doesn't. Creating it."
-        $creationParams = Remove-M365DSCAuthenticationParameter -BoundParameters $PSBoundParameters
+        $creationParams = Rename-M365DSCCimInstanceParameter -Properties $boundParameters
         $creationParams.Remove('IsSingleInstance') | Out-Null
 
         $emails = ''
-        foreach ($email in $creationParams.AlternateNotificationEmails)
+        foreach ($email in $creationParams.alternateNotificationEmails)
         {
             $emails += $email + ';'
         }
         $emails = $emails.TrimEnd(';')
-        $creationParams.AlternateNotificationEmails = $emails
+        $creationParams.alternateNotificationEmails = $emails
         New-MgGroupLifecyclePolicy -BodyParameter $creationParams
     }
     elseif ($Ensure -eq 'Present' -and $currentPolicy.Ensure -eq 'Present')
     {
-        $updateParams = Remove-M365DSCAuthenticationParameter -BoundParameters $PSBoundParameters
+        $updateParams = Rename-M365DSCCimInstanceParameter -Properties $boundParameters
         $updateParams.Remove('IsSingleInstance') | Out-Null
 
         $emails = ''
-        foreach ($email in $updateParams.AlternateNotificationEmails)
+        foreach ($email in $updateParams.alternateNotificationEmails)
         {
             $emails += $email + ';'
         }
         $emails = $emails.TrimEnd(';')
-        $updateParams.AlternateNotificationEmails = $emails
+        $updateParams.alternateNotificationEmails = $emails
 
         Write-Verbose -Message "The Group Lifecycle Policy exists but it's not in the Desired State. Updating it."
         Update-MgGroupLifecyclePolicy -GroupLifecyclePolicyId (Get-MgGroupLifecyclePolicy).Id -BodyParameter $updateParams